requirements.txt
python==3.6.0
Django==3.0.7
djangorestframework==3.11.0

